+++ include/prompt.h
@@ -3,6 +3,6 @@
 
 #include <string>
 
-std::string get_prompt_string(bool end_of_file_in = false);
+std::string get_prompt_string();
 
 #endif // PROMPT_H